Refactor common pieces of single- and multi-headed data_vault
